refactor: rename commands
