--- messages/kc.apex-factory.trigger-framework.trigger.md
@@ -1,19 +1,0 @@
-# summary
-
-Summary of a command.
-
-# description
-
-More information about a command. Don't repeat the summary.
-
-# examples
-
-- <%= config.bin %> <%= command.id %>
-
-# flags.target-dir.summary
-
-The target directory for your apex classes and triggers (default=force-app/main/default). There should be a classes/ and triggers/ directory at this location.
-
-# flags.sobject.summary
-
-The name of the SObject that the trigger will be created for.
